% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik.R
\name{log_lik}
\alias{log_lik}
\title{Log Likelihood per Point and based on Posterior Mean}
\usage{
log_lik(
  samples,
  data,
  Ngrid,
  lv_mu,
  lv_cov,
  log_joint_i,
  parallel = TRUE,
  n_cores = detectCores() - 2,
  packages = c("matrixStats", "statmod", "mvtnorm", "extraDistr")
)
}
\arguments{
\item{samples}{A matrix or data frame of Bayesian posterior samples of model parameters.}

\item{data}{A list of data, including an element 'N' which indicates the number of units}

\item{Ngrid}{Number of grid per dimension.}

\item{lv_mu}{A list of posterior means for the latent variables.}

\item{lv_cov}{A list of posterior covariance matrix for latent variables.}

\item{log_joint_i}{A function that computes the log joint probability for unit i.}

\item{parallel}{A logical indicating whether to compute in parallel (default is TRUE).}

\item{n_cores}{Number of cores to use for parallel computation. Defaults to \code{detectCores() - 2}.}

\item{packages}{A character vector of package names to be loaded in the parallel environment.}
}
\value{
A list containing two elements:
\item{loglik_point}{A matrix of log likelihoods for each data point.}
\item{loglik_postmean}{A vector of log likelihoods computed using the posterior mean of the parameters.}
}
\description{
Computes ...
}
