% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_i.R
\name{log_lik_i}
\alias{log_lik_i}
\title{Log Likelihood for Unit i}
\usage{
log_lik_i(samples_s, data, i, Ngrid, lv_mu, lv_cov, log_joint_i)
}
\arguments{
\item{samples_s}{A vector of Bayesian posterior sample s of model parameters.}

\item{data}{A list of data, including an element 'N' which indicates the number of units}

\item{i}{Index of the unit for which to compute the log likelihood.}

\item{Ngrid}{Number of grid per dimension.}

\item{lv_mu}{A list of posterior means for the latent variables.}

\item{lv_cov}{A list of posterior covariance matrix for latent variables.}

\item{log_joint_i}{A function that computes the log joint probability for unit i.}
}
\value{
The log likelihood for unit i.
}
\description{
Log Likelihood for Unit i
}
