% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_marglik.R
\name{log_marglik}
\alias{log_marglik}
\title{Log Marginal Likelihood per Point and based on Posterior Mean}
\usage{
log_marglik(
  samples,
  data,
  Ngrid,
  lv_mu,
  lv_cov,
  log_joint_i,
  parallel = TRUE,
  n_cores = detectCores() - 2,
  packages = c("matrixStats", "statmod", "mvtnorm", "extraDistr", "truncdist")
)
}
\arguments{
\item{samples}{A matrix or data frame of Bayesian posterior samples of model parameters.
Each row represents a sample, and each column represents a parameter.}

\item{data}{A list of data, including an element 'N' which indicates the number of units.}

\item{Ngrid}{Number of grid points (quadrature nodes) per dimension.}

\item{lv_mu}{A list of posterior means for the latent variables.
Each element corresponds to the posterior mean of the latent variables
for a specific unit.}

\item{lv_cov}{A list of posterior covariance matrix for the latent variables.
Each element corresponds to the posterior covariance matrix of
the latent variables for a specific unit.}

\item{log_joint_i}{A user-defined function that computes the log joint density
for a given unit. This function should take the following arguments:
\itemize{
\item \code{samples_s}: A vector of parameter values from a posterior sample.
\item \code{data}: The data list.
\item \code{i}: The index of the unit.
\item \code{Ngrid}: Number of grid points (quadrature nodes) per dimension.
\item \code{nodes}: A matrix of quadrature nodes transformed using the latent variable mean and covariance.
}}

\item{parallel}{A logical indicating whether to compute in parallel (default is TRUE).}

\item{n_cores}{Number of cores to use for parallel computation. Defaults to \code{detectCores() - 2}.}

\item{packages}{A character vector of package names to be loaded in the parallel environment.}
}
\value{
A list containing two elements:
\item{log_marglik_point}{A matrix of log marginal likelihoods for each data point.
Each row corresponds to a posterior sample, and each column corresponds to a unit.}
\item{log_marglik_postmean}{A vector of log marginal likelihoods computed using the posterior mean
of the parameters. Each element corresponds to a unit.}
}
\description{
This function computes the log marginal likelihood for each data point and
the log marginal likelihood based on the posterior mean of the parameters.
It is designed for Bayesian latent variable models, where the marginal likelihood
is computed by integrating out latent variables using numerical quadrature.
The function supports parallel computation to improve efficiency for large datasets.
}
