% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_marg_lik.R
\name{log_marg_lik}
\alias{log_marg_lik}
\title{Log Marginal Likelihood for the Model}
\usage{
log_marg_lik(
  samples,
  data,
  Ngrid,
  lv_mu,
  lv_cov,
  log_joint_i,
  log_prior,
  lb,
  ub,
  ...
)
}
\arguments{
\item{samples}{A matrix or data frame of Bayesian posterior samples of model parameters.}

\item{data}{A list of data, including an element 'N' which indicates the number of persons.}

\item{Ngrid}{Number of grid per dimension.}

\item{lv_mu}{A list of posterior means for the latent variables.}

\item{lv_cov}{A list of posterior covariance matrix for latent variables.}

\item{log_joint_i}{A function that computes the log joint probability for unit i.}

\item{log_prior}{A function that computes the log prior probabilities of the model parameters.}

\item{lb}{named vector with lower bounds for parameters. This is required for
the \code{\link[bridgesampling]{bridge_sampler}} function and defines the
minimum value for the parameter space over which the marginal likelihood is computed.}

\item{ub}{named vector with upper bounds for parameters. This is required for
the \code{\link[bridgesampling]{bridge_sampler}} function and defines the
maximum value for the parameter space over which the marginal likelihood is computed.}

\item{...}{Additional arguments passed directly to the \code{bridge_sampler} function.}
}
\value{
A list containing the result of the bridge sampling, which includes the log marginal
likelihood estimate and other information related to the bridge sampling process.
The return value is typically an object from the \code{bridge_sampler} function containing:
\itemize{
\item \code{logml}: estimate of log marginal likelihood.
\item \code{niter}: number of iterations of the iterative updating scheme.
}
}
\description{
Computes the log marginal likelihood of the Bayesian latent variable models
using bridge sampling.
The function takes posterior samples of the model parameters, a log-likelihood function,
and a log-prior function, and then uses the \code{bridge_sampler} function from the \code{bridgesampling}
package to estimate the log marginal likelihood.
}
\examples{
\donttest{
log_marg_lik(...)  # need another file
}

}
